Update the sample to use AdalJS v1.0.15.
This version adds 2 new APIs, acquireTokenRedirect and acquireTokenPopup allowing to acquire token interactively with the help of a UI. In previous versions of Adal.JS, we only supported the acquireToken API, which used a hidden iframe to acquire a token for a resource. This AcquireToken API would fail silently if the cookies are not sent with the request or the session is expired with AAD. You can now call acquireTokenRedirect/acquireTokenPopup on this failure which would show a UI where users can enter their credentials and renew their session with AAD to get an access token.

Updated the todoListCtrl.js to react to the claim challenge.